pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arc-solver-unified"
version = "1.0.0"
description = "ARC-AGI Unified Solver: Multi-Provider LLMs + MCTS + Critic Meta-Reasoning"
requires-python = ">=3.11"
dependencies = [
    # API Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",

    # Data & Validation
    "pydantic>=2.7.0",
    "numpy>=1.26.0",
    "scipy>=1.11.0",

    # HTTP Clients
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",

    # Config & Environment
    "PyYAML>=6.0",
    "python-dotenv>=1.0.0",

    # Utilities
    "tenacity>=8.2.0",
    "rich>=13.0.0",
    "typer>=0.9.0",

    # Logging
    "loguru>=0.7.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["arc_solver*"]
exclude = ["tests*", "configs*"]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black",
    "isort",
    "mypy",
]

[project.scripts]
arc-solve = "arc_solver.cli:app"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
asyncio_mode = "auto"
